fix: s/SplitPackageDetailts/SplitPackageDetails/
