+++ src/components/SplitPackageDetails.tsx
